backend/requirements.txt
alibabacloud_ocr_api20210707==3.1.3
alibabacloud_tea_openapi==0.3.15
alibabacloud_tea_util==0.3.13
numpy<2
cnocr==2.3.1
Flask==3.1.1
Flask_Cors==5.0.0
flask_sqlalchemy==3.1.1
ImageHash==4.3.2
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.60
langchain_openai==0.3.17
openai_whisper==20240930
opencv_python_headless==4.9.0.80
pandas==2.2.3
Pillow==11.2.1
psutil==5.9.8
pydantic==2.11.4
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.1.0
scikit-image
SQLAlchemy==2.0.37
Werkzeug==3.1.3
jwt
tencentcloud-sdk-python-ocr
rank_bm25